+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/release.ftl
@@ -1,4 +1,4 @@
-<#assign releaseModels = dataFactory.newReleaseModels() />
+<#assign releaseModels = releaseDataFactory.newReleaseModels() />
 
 <#list releaseModels as releaseModel>
 	${dataFactory.toInsertSQL(releaseModel)}
